fix(nlp): fix rul_days type from int to float

+++ nlp/api/schemas.py
@@ -93,7 +93,7 @@ class LiveContextSnapshot(BaseModel):
     pressure_psi  : Optional[float] = None
     rpm           : Optional[float] = None
     ml_prediction : str = "unknown"
-    rul_days      : Optional[int] = None
+    rul_days      : Optional[float] = None
     active_alerts : List[str] = []
     data_source   : str = "mock"
